+++ engine/runtime-compiler/src/main/scala/org/enso/compiler/pass/resolve/TypeNames.scala
@@ -45,7 +45,7 @@ case object TypeNames extends IRPass {
   ): Module = {
     val bindingsMap =
       ir.unsafeGetMetadata(BindingAnalysis, "bindings analysis did not run")
-    ir.copyWithBindings(bindings = ir.bindings.map { d =>
+    ir.copyWithBindings(ir.bindings.map { d =>
       val selfTypeInfo: SelfTypeInfo = d match {
         case t: Definition.Type => SelfTypeInfo.fromTypeDefinition(t)
         case m: Method.Explicit =>
@@ -216,13 +216,14 @@ case object TypeNames extends IRPass {
       })
       .fold(
         error =>
-          errors.Resolution(name, errors.Resolution.ResolverError(error)),
+          errors.Resolution
+            .create(name, new errors.Resolution.ResolverError(error)),
         n =>
           n.getMetadata(this).get.target match {
             case _: ResolvedModule =>
-              errors.Resolution(
+              errors.Resolution.create(
                 n,
-                errors.Resolution.UnexpectedModule("type signature")
+                new errors.Resolution.UnexpectedModule("type signature")
               )
             case _ => n
           }